revert ring index type change

+++ pkg/queue/ring.go
@@ -14,9 +14,9 @@ func NewRing[T any](n int) (r Ring[T]) {
 
 // Ring ...
 type Ring[T any] struct {
-	mask uint64
-	low  uint64
-	high uint64
+	mask int
+	low  int
+	high int
 	v    []T
 }
 
@@ -27,7 +27,7 @@ func (r *Ring[T]) Cap() int {
 
 // Len ...
 func (r *Ring[T]) Len() int {
-	return int(r.high - r.low)
+	return r.high - r.low
 }
 
 // Resize ...
@@ -43,7 +43,7 @@ func (r *Ring[T]) Resize(n int) {
 	}
 
 	v := make([]T, n)
-	mask := uint64(n - 1)
+	mask := n - 1
 
 	vi := r.low & mask
 	i := r.low & r.mask
@@ -119,7 +119,7 @@ func (r *Ring[T]) Iterator() *RingIterator[T] {
 
 // RingIterator ...
 type RingIterator[T any] struct {
-	i uint64
+	i int
 	r *Ring[T]
 }
